requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.3.0
opencv-python>=4.5.0
tqdm>=4.62.0
tensorboard>=2.7.0
matplotlib>=3.4.0
scikit-image>=0.18.0
